Integrated data and utils into TreeGridAxis.
