--- js/parts-gantt/TreeGridData.js
@@ -1,153 +1,0 @@
-/* *
- *
- *  (c) 2016 Highsoft AS
- *  Authors: Jon Arild Nygard
- *
- *  License: www.highcharts.com/license
- *
- *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
- *
- * */
-'use strict';
-import Tree from './Tree.js';
-import U from '../parts/Utilities.js';
-var find = U.find, isObject = U.isObject, isString = U.isString;
-/* eslint-disable valid-jsdoc */
-/**
- * Creates a tree structure of the data, and the treegrid. Calculates
- * categories, and y-values of points based on the tree.
- *
- * @private
- * @function getTreeGridFromData
- *
- * @param {Array<Highcharts.GanttPointOptions>} data
- * All the data points to display in the axis.
- *
- * @param {boolean} uniqueNames
- * Wether or not the data node with the same name should share grid cell. If
- * true they do share cell. False by default.
- *
- * @param {number} numberOfSeries
- *
- * @return {object}
- * Returns an object containing categories, mapOfIdToNode,
- * mapOfPosToGridNode, and tree.
- *
- * @todo There should be only one point per line.
- * @todo It should be optional to have one category per point, or merge
- *       cells
- * @todo Add unit-tests.
- */
-function getTreeGridFromData(data, uniqueNames, numberOfSeries) {
-    var categories = [], collapsedNodes = [], mapOfIdToNode = {}, mapOfPosToGridNode = {}, posIterator = -1, uniqueNamesEnabled = typeof uniqueNames === 'boolean' ? uniqueNames : false, tree;
-    // Build the tree from the series data.
-    var treeParams = {
-        // After the children has been created.
-        after: function (node) {
-            var gridNode = mapOfPosToGridNode[node.pos], height = 0, descendants = 0;
-            gridNode.children.forEach(function (child) {
-                descendants += (child.descendants || 0) + 1;
-                height = Math.max((child.height || 0) + 1, height);
-            });
-            gridNode.descendants = descendants;
-            gridNode.height = height;
-            if (gridNode.collapsed) {
-                collapsedNodes.push(gridNode);
-            }
-        },
-        // Before the children has been created.
-        before: function (node) {
-            var data = isObject(node.data, true) ? node.data : {}, name = isString(data.name) ? data.name : '', parentNode = mapOfIdToNode[node.parent], parentGridNode = (isObject(parentNode, true) ?
-                mapOfPosToGridNode[parentNode.pos] :
-                null), hasSameName = function (x) {
-                return x.name === name;
-            }, gridNode, pos;
-            // If not unique names, look for sibling node with the same name
-            if (uniqueNamesEnabled &&
-                isObject(parentGridNode, true) &&
-                !!(gridNode = find(parentGridNode.children, hasSameName))) {
-                // If there is a gridNode with the same name, reuse position
-                pos = gridNode.pos;
-                // Add data node to list of nodes in the grid node.
-                gridNode.nodes.push(node);
-            }
-            else {
-                // If it is a new grid node, increment position.
-                pos = posIterator++;
-            }
-            // Add new grid node to map.
-            if (!mapOfPosToGridNode[pos]) {
-                mapOfPosToGridNode[pos] = gridNode = {
-                    depth: parentGridNode ? parentGridNode.depth + 1 : 0,
-                    name: name,
-                    nodes: [node],
-                    children: [],
-                    pos: pos
-                };
-                // If not root, then add name to categories.
-                if (pos !== -1) {
-                    categories.push(name);
-                }
-                // Add name to list of children.
-                if (isObject(parentGridNode, true)) {
-                    parentGridNode.children.push(gridNode);
-                }
-            }
-            // Add data node to map
-            if (isString(node.id)) {
-                mapOfIdToNode[node.id] = node;
-            }
-            // If one of the points are collapsed, then start the grid node
-            // in collapsed state.
-            if (gridNode &&
-                data.collapsed === true) {
-                gridNode.collapsed = true;
-            }
-            // Assign pos to data node
-            node.pos = pos;
-        }
-    };
-    var updateYValuesAndTickPos = function (map, numberOfSeries) {
-        var setValues = function (gridNode, start, result) {
-            var nodes = gridNode.nodes, end = start + (start === -1 ? 0 : numberOfSeries - 1), diff = (end - start) / 2, padding = 0.5, pos = start + diff;
-            nodes.forEach(function (node) {
-                var data = node.data;
-                if (isObject(data, true)) {
-                    // Update point
-                    data.y = start + (data.seriesIndex || 0);
-                    // Remove the property once used
-                    delete data.seriesIndex;
-                }
-                node.pos = pos;
-            });
-            result[pos] = gridNode;
-            gridNode.pos = pos;
-            gridNode.tickmarkOffset = diff + padding;
-            gridNode.collapseStart = end + padding;
-            gridNode.children.forEach(function (child) {
-                setValues(child, end + 1, result);
-                end = (child.collapseEnd || 0) - padding;
-            });
-            // Set collapseEnd to the end of the last child node.
-            gridNode.collapseEnd = end + padding;
-            return result;
-        };
-        return setValues(map['-1'], -1, {});
-    };
-    // Create tree from data
-    tree = Tree.getTree(data, treeParams);
-    // Update y values of data, and set calculate tick positions.
-    mapOfPosToGridNode = updateYValuesAndTickPos(mapOfPosToGridNode, numberOfSeries);
-    // Return the resulting data.
-    return {
-        categories: categories,
-        mapOfIdToNode: mapOfIdToNode,
-        mapOfPosToGridNode: mapOfPosToGridNode,
-        collapsedNodes: collapsedNodes,
-        tree: tree
-    };
-}
-var TreeGridData = {
-    getTreeGridFromData: getTreeGridFromData
-};
-export default TreeGridData;
